Task 4: Added the useParams and tested the new route of gamedetails and modules exported to gameinfo.js
