+++ staff/gustavo-rojas/course-3/task-4/site-nysl/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import logo from './nysl_logo.svg';
 import './App.css';
 import {nysl_league,sport_events,page_home_header,logo_alttext,logo_width} from "./components/home.js"
-import {game_info,additional_info_game,table_games_header,page_gameinfo_header,dayweek_warning,game_locations,GameDetails } from "./components/gameinfo.js"
+import {Gamedetails,Gameinfo} from "./components/gameinfo.js"
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {faHome,faInfoCircle,faRuler,faEnvelope,faFutbol,faUserPlus} from '@fortawesome/free-solid-svg-icons';
 import {
@@ -11,20 +11,20 @@ import {
   Route,
   NavLink
 } from "react-router-dom";
-import validator from 'validator'
+
 export default function App() {
   return (
     <Router>
       <div>
-        <Switch>
-        <Route path="/contact">    <Contact />  </Route>
-        <Route path="/regform">    <Regform />  </Route>
-        <Route path="/rules">      <Rules />    </Route>
-        <Route path="/about">      <About />    </Route>
-        <Route path="/gameinfo">   <Gameinfo /> </Route>
-        <Route path="/">           <Home />     </Route>
-        <Route path="/gamedetails/:id">         <GameDetails /> </Route>
-        </Switch>
+      <Switch>
+        <Route path="/contact">                 <Contact />     </Route>
+        <Route path="/regform">                 <Regform />     </Route>
+        <Route path="/rules">                   <Rules />       </Route>
+        <Route path="/about">                   <About />       </Route>
+        <Route path="/gameinfo">                <Gameinfo />    </Route>
+        <Route path="/gamedetails/:id">         <Gamedetails /> </Route>
+        <Route path="/">                        <Home />        </Route>
+      </Switch>
       </div>
       <nav className="navbar fixed-bottom navbar-light bg-light" >
               <NavLink to="/"> <FontAwesomeIcon size="lg" icon={faHome}/> </NavLink>
@@ -33,7 +33,7 @@ export default function App() {
               <NavLink to="/contact"><FontAwesomeIcon size="lg" icon={faEnvelope}/></NavLink>
               <NavLink to="/regform"><FontAwesomeIcon size="lg" icon={faUserPlus}/></NavLink>
               <NavLink to="/about"><FontAwesomeIcon size="lg" icon={faInfoCircle}/> </NavLink>
-     </nav>
+        </nav>
     </Router>
   );
 }
@@ -60,61 +60,15 @@ function About() {
   </div>
 }
 
-function Gameinfo() {
-  return  <div>
-  <h5 > <img src={logo} alt={logo_alttext} width={logo_width} /> { nysl_league.title }</h5>
-  <h5>{page_gameinfo_header}</h5>
-  <h5>{dayweek_warning}</h5>
-  <table className="table">
-  <thead>
-  {table_games_header.map(header => {
-        return (
-        <tr key={header.id}><th>{header.first_col}</th><th>{header.second_col}</th><th>{header.third_col}</th><th>{header.fourth_col}</th>
-        </tr>
-        )
-      })}
-  </thead>
-  <tbody>
-      {game_info.map(game => {
-        var location_temp=game.Location;
-        var game_location_temp=game_locations[location_temp][0].name_location;
-        return (
-        <tr key={game.id}><td>{game.Date}</td><td>{game.Teams}</td><td>{game_location_temp}</td><td>{game.Times}</td>
-        </tr>
-        )
-      })}
-   </tbody>
-    </table>
 
-   <table className="table">
-   <tbody>
-      {additional_info_game.map(info => {
-        var is_email=validator.isEmail(info.info);
-        var info_temp ="N.A";
-        if(is_email) 
-          {
-            info_temp= <a href={`mailto: ${info.info}`}>{info.info}</a>
-          } 
-          else {
-            info_temp=info.info
-          };
-        return (
-        <tr key={info.id}>
-          <td>{info.label}</td>
-          <td>{info_temp}</td></tr>
-        )
-      })}
-      </tbody> 
-        </table>
 
-  </div>
-}
+
 
 
 function Rules() {
   return  <div>
   <h5 > <img src={logo} alt={logo_alttext} width={logo_width} /> { nysl_league.title }</h5>
-  <h5>Rules and Policies</h5>
+  <h5>Rules and Policies </h5>
   </div>
 }
 
@@ -131,3 +85,4 @@ function Contact() {
   <h5>Contact</h5>
   </div>
 }
+
